base/pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "1.0.0"
description = "FastAPI Template"
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.128.0,<0.129.0"
uvicorn = { version = ">=0.40.0,<0.41.0", extras = ["standard"] }
sqlalchemy = ">=2.0.45,<3.0.0"
aiomysql = ">=0.2.0,<0.3.0"
python-jose = { version = ">=3.5.0,<4.0.0", extras = ["cryptography"] }
pydantic = { version = ">=2.12.5,<3.0.0", extras = ["email"] }
pydantic-settings = ">=2.8.0,<3.0.0"
sqlmodel = ">=0.0.31,<0.0.32"
python-multipart = ">=0.0.21,<0.0.22"
types-python-jose = ">=3.5.0.20250531,<4.0.0.0"
bcrypt = ">=5.0.0,<6.0.0"
platformdirs = "^4.5.1"
alembic = ">=1.15.0,<2.0.0"
jinja2 = "^3.1.6"

[dependency-groups]
dev = [
    "pre-commit (>=4.3.0,<5.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "autopep8 (>=2.3.2,<3.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "flake8-bugbear (>=24.12.12,<25.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "types-tabulate (>=0.9.0.20241207,<0.10.0.0)",
    "isort (>=7.0.0,<8.0.0)",
]

[tool.flake8]
extend-immutable-calls = ["typer.Option", "Depends"]

[tool.mypy]
plugins = ["pydantic.mypy"]